+++ Changelog.md
@@ -1,2 +1,2 @@
 # Changelog
-- Allow pawns with AnimalInsectSpace trade tag to survive space
+- Add patch to remove any light overlay in vacuum utility